+++ classes/gui/block/RichTextBlock/RichTextBlockPresentationView.php
@@ -118,6 +118,7 @@ final class RichTextBlockPresentationView implements Renderable {
 			$outerTemplate->setVariable('SEQUENCE_INPUT', $input->render());
 		}
 		$outerTemplate->setVariable('CONTENT', $template->get());
+		$outerTemplate->setVariable('SEQUENCE', $this->model->getSequence());
 		return $outerTemplate;
 	}
 }