+++ classes/gui/settings/SettingEditFormView.php
@@ -100,7 +100,7 @@ final class SettingEditFormView extends ilPropertyFormGUI {
 		$visibilitySectionHeader->setTitle($this->plugin->txt('common_visibility'));
 		$this->addItem($visibilitySectionHeader);
 
-		$radioGroup = new ilRadioGroupInputGUI($this->plugin->txt('visibility_title'), self::POST_DEFAULT_VISIBILITY);
+		$radioGroup = new ilRadioGroupInputGUI($this->plugin->txt('setting_visibility_default'), self::POST_DEFAULT_VISIBILITY);
 		$radioGroup->addOption(new ilRadioOption($this->plugin->txt('visibility_always'), Visibility::ALWAYS));
 		$radioGroup->addOption(new ilRadioOption($this->plugin->txt('visibility_after_visit_place'), Visibility::AFTER_VISIT_PLACE));
 		$radioGroup->addOption(new ilRadioOption($this->plugin->txt('visibility_only_at_place'), Visibility::ONLY_AT_PLACE));
@@ -146,6 +146,7 @@ final class SettingEditFormView extends ilPropertyFormGUI {
 		$this->configuration->setDefaultVisibility($visibility);
 		$this->configuration->setLongitude(doubleval($this->getInput(self::POST_LOCATION)['longitude']));
 		$this->configuration->setLatitude(doubleval($this->getInput(self::POST_LOCATION)['latitude']));
+		$this->configuration->setMapZoom(intval($this->getInput(self::POST_LOCATION)['zoom']));
 		$this->configuration->setOnline(intval($this->getInput(self::POST_ONLINE)) === 1);
 		$this->configuration->setRadius(intval($this->getInput(self::POST_LOCATION_RADIUS)));
 		$this->configuration->setElevation(0);
@@ -164,7 +165,11 @@ final class SettingEditFormView extends ilPropertyFormGUI {
 		$values = [
 			self::POST_DEFAULT_VISIBILITY   => $this->configuration->getDefaultVisibility(),
 			self::POST_ONLINE               => $this->configuration->isOnline(),
-			self::POST_LOCATION             => ['latitude' => $this->configuration->getLatitude(), 'longitude' => $this->configuration->getLongitude()],
+			self::POST_LOCATION             => [
+				'latitude'  => $this->configuration->getLatitude(),
+				'longitude' => $this->configuration->getLongitude(),
+				'zoom'      => $this->configuration->getMapZoom()
+			],
 			self::POST_LOCATION_RADIUS      => $this->configuration->getRadius(),
 			self::POST_TITLE                => $this->configuration->getTitle(),
 			self::POST_DESCRIPTION          =>$this->configuration->getDescription(),