+++ CHANGELOG.md
@@ -4,15 +4,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0]
+### Added
+- Rich text blocks.
+- Picture blocks.
+- Accordion blocks.
+- Map
+- Video blocks.
+- Link blocks.
+- Visibilities for blocks
+	- Always
+	- Only at place
+	- After visit place
+	- Never
+- Configurable default visibility for blocks
+- Configurable map zoom level.
+
+
+
 ## [Unreleased]
 ### Added
-- composer configuration
-- License
-- Changelog
-- gitignore
 ### Changed
 ### Deprecated
 ### Removed
 ### Fixed
 ### Security
-